pyproject.toml
[project]
name = "trustgraph-finance-copilot"
version = "0.1.0"
description = "Verifiable Finance Copilot with executable citations (RAG over SEC) — Groq + uv + DuckDB + FAISS"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "python-dotenv>=1.0.1",
  "streamlit>=1.39.0",
  "requests>=2.32.3",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.3.0",
  "pandas>=2.2.2",
  "numpy>=1.26.4",
  "duckdb>=1.1.2",
  "faiss-cpu>=1.8.0.post1",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.1.1",
  "scikit-learn>=1.5.2",
  "tqdm>=4.66.6",
  "pydantic>=2.9.2",
  "sec-edgar-downloader>=5.0.2",
  "yfinance>=0.2.43",
  "groq>=0.11.0",
]

[tool.uv]
index-strategy = "first-index"
dev-dependencies = []

